server_rest_api/requirements_linux.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==3.4.7
Deprecated==1.2.12
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-oauth-toolkit==1.5.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==3.2
jwcrypto==1.0
oauthlib==3.1.1
pkg_resources==0.0.0
pycparser==2.20
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
wrapt==1.12.1
